Support abstract namespace for unix domain sockets
On Linux, sockaddr_un.sun_path starting with a null byte places the socket into
the abstract namespace. A common convention (systemd, upstart) to configure
this is to start the path with @.

This avoids file system accesses for linking/unlinking on the server,
and opening on clients, and is useful in chroot environments.

Signed-off-by: Niklas Fiekas <niklas.fiekas@sartorius.com>